Use lowercase headers on Rack 3
The Rack SPEC is changing in Rack 3 to not allow uppercase
ASCII in response headers (it's also changing to require
response headers to be an unfrozen hash).  This change
should allow rack_csrf to follow the Rack 3 SPEC.  For
backwards compatibility, it does not change the behavior when
used with older versions of Rack.

This checks for Rack 2.3 instead of Rack 3, since Rack that's
the current Rack main branch version.  However, the next Rack
version will be Rack 3.0, not Rack 2.3.

An alternative approach to this that would be more compatible
would be to use Rack::Headers if available (this has been added
to the Rack main branch, and will ship with Rack 3).  This allows
for case-insensitive access to the headers.