When using Rack 3, only load rack/request and rack/utils
These are the only parts of Rack that rack_csrf uses.  Starting in
Rack 3, it is recommended that users of Rack load the parts of
Rack they are using, instead of requiring rack itself and relying
on the autoloads it adds.

In Rack <3, still load rack itself, since loading parts of Rack
was not guaranteed to work in earlier versions.

Load rack/mock when testing, since the tests use Rack::MockRequest.